Update index.js
F

+++ index.js
@@ -20,9 +20,12 @@ msg.channel.send('bot version 1.0 by MrBoom#4863');
 
 client.on('message', msg => {
 if (msg.content === prefix+'restart') {
+if (message.author.id === '556160607339020315') {
 channel.send('Resetting...')
     .then(msg => client.destroy())
-    .then(() => client.login
+    .then(() => client.login('NjQzMDQ5Mzg4NzY3MTgyODQ5.Xcf0aQ.gMllMX1waEivtAprlNME3q5YD5c');
+} else {
+msg.channel.send('You not bot creator!');
 }
 }
